+++ ratelimit_test.go
@@ -1,7 +1,6 @@
 package ratelimit
 
 import (
-	"fmt"
 	"sync"
 	"testing"
 	"time"
@@ -39,7 +38,7 @@ type runnerImpl struct {
 }
 
 func runTest(t *testing.T, fn func(testRunner)) {
-	implementations := []struct {
+	impls := []struct {
 		name        string
 		constructor func(int, ...Option) Limiter
 	}{
@@ -57,7 +56,7 @@ func runTest(t *testing.T, fn func(testRunner)) {
 		},
 	}
 
-	for _, tt := range implementations {
+	for _, tt := range impls {
 		t.Run(tt.name, func(t *testing.T) {
 			r := runnerImpl{
 				t:           t,
@@ -131,30 +130,6 @@ func (r *runnerImpl) goWait(fn func()) {
 		fn()
 	}()
 	wg.Wait()
-}
-
-func Example() {
-	rl := New(100) // per second
-
-	prev := time.Now()
-	for i := 0; i < 10; i++ {
-		now := rl.Take()
-		if i > 0 {
-			fmt.Println(i, now.Sub(prev))
-		}
-		prev = now
-	}
-
-	// Output:
-	// 1 10ms
-	// 2 10ms
-	// 3 10ms
-	// 4 10ms
-	// 5 10ms
-	// 6 10ms
-	// 7 10ms
-	// 8 10ms
-	// 9 10ms
 }
 
 func TestUnlimited(t *testing.T) {
